kalimat-cash-main-backend/backend/requirements.txt
gunicorn
djangorestframework
celery
django-jazzmin
django-celery-results
django-celery-beat
psycopg2-binary
django-import-export

